pyproject.toml
[tool.poetry]
name = "backend-project"
version = "0.1.0"
description = "electronic library"
authors = ["Vlad Verevkin, Kirill Ivantsov, Daniel Kotov"]
readme = "README.md"
packages = [{include = "project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "3.2.16"
djangorestframework = "3.14.0"
django-environ = "0.8.0"
djoser = {version = "2.1.0"}
celery = {version = "5.2.7", extras= ["redis"]}
jsonschema = "^4.23.0"
#psycopg2 = "2.9.10"
loguru = "0.7.3"

alembic = "1.14.0"
asyncpg = "0.30.0"
sqlalchemy = "2.0.36"
pydantic-settings = "2.7.0"

fastapi = "^0.100.0"
uvicorn = "^0.22.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pydantic = {extras = ["email"], version = "^2.10.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
